requirements.txt
fastapi
uvicorn
redis
pydantic
pydantic-settings
google-generativeai
httpx
python-dotenv
aiohttp
duckduckgo-search
beautifulsoup4
